Added Eligibility Criteria Engine

+++ backend/src/main/java/com/example/gov_scheme_backend/repositories/SchemeRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SchemeRepo extends JpaRepository<Schemes,Integer> {
+public interface SchemeRepo extends JpaRepository<Schemes,Long> {
 
 }
